+++ src/Models/Constants/Types.ts
@@ -15,3 +15,23 @@ export type AdapterState<T extends IAdapterData> = {
     isLoading: boolean;
     isLongPolling: boolean;
 };
+
+export type KeyValuePairData = Record<string, any>;
+
+export type TsiClientData = any[];
+
+export type CancellablePromise<T> = {
+    /** Wrapped promise - throws CancelledPromiseError if cancelled */
+    promise: Promise<T>;
+    /** Function to cancel promise */
+    cancel: () => void;
+};
+
+export type UseLongPollParams = {
+    /** Function to long poll at set interval */
+    callback: () => Promise<any>;
+    /** Time between each callback execution */
+    pollingIntervalMillis?: number;
+    /** Length of time UI pulse state remains true after callback completion - use to indicate updated data */
+    pulseTimeoutMillis?: number;
+};
